Updated Quotes rule

+++ source/eslintrc.ts
@@ -1701,7 +1701,7 @@ module.exports = {
     ],
 
     // specify whether double or single quotes should be used
-    quotes: ['error', 'single', { avoidEscape: true }],
+    quotes: ['error', 'double', { avoidEscape: true }],
 
     // do not require jsdoc
     // https://eslint.org/docs/rules/require-jsdoc
